Changing the packing strategy from being non-blocking but potentially redundant, to being non-redundant but potentially blocking.
So far, when multiple threads needed to pack a LHS or RHS block that hadn't been packed yet, both could concurrently do the packing work. This was OK as any concurrent writes would be writing the same values, but was still the reason why packing code required relaxed atomic stores. The upside was that our packing code never blocked.

Requiring relaxed atomic stores was a downside in its own right, already from a generic source code complexity perspective, and more so in the context of architecture-specific packing code paths that might need to use SIMD intrinsics, for which there might be no relaxed-atomic version. Work-arounds were available but further added to the source-code-complexity concern.

Meanwhile, a naive blocking version of this code was found to perform just as well in most cases and only slightly worse in a worst-case of multithreading across CPU cores of very different speeds.

This CL goes a little farther than just a naive blocking version. It also takes advantage of the observation that if a thread finds that it needs to pack a LHS block that is already being packed by another thread, then instead of just waiting for that, it can instead be packing the RHS block, and conversely. This should reduce the amount of time spent waiting.

A further CL is expected to go farther in letting threads do other packing work, e.g. moving on to the next few blocks, which would typically be used in the near future.

Regardless of specifics, the fact that we have these different tricks to play to reduce overhead as needed, makes me confident that this potentially blocking approach is the right one.

PiperOrigin-RevId: 259994760