Use preload-for-write instructions before actual store instructions in kernels.
Introduce a RUY_PREFETCH_STORE optset, separate from RUY_PREFETCH_LOAD. Unlike RUY_PREFETCH_LOAD which is detrimental in some contexts (matmul-only benchmarking), RUY_PREFETCH_STORE seems never detrimental, so we need to control it separately.

This is a substantial speedup when the destination matrix stride is close to a multiple of L1 cache aliasing periodicity. For example, on a typical ARM CPU that is 1024-byte periodicity, so for float matrices that happens whenever the destination matrix has a number of rows close to 1024/sizeof(float) = 256.  The impact varies gradually as one gets closer to such values. Typically, the impact is large when one is within 64 bytes (1 cache line size) of the nearest such value.

For shallow shapes (small depth, e.g. depth=16 as is common in MobileNet v3) this can be a 2x speedup, as there is not much arithmetic to amortize the high cost of these writes.

As far as I understand, what we really want here is a 'non-temporal store instruction' such as the STNP instruction but I'm not aware of a NEON non-temporal instruction. Using a prefetch-stream instruction just before a regular store seems to convey a strictly weaker hint (we are not conveying that the order of observation does not matter) but already seems to be effective at avoiding cache aliasing issues.

I also experimented with 'pstl3strm' instead of 'pstl1strm' but that wasn't better.

PiperOrigin-RevId: 291440490