Make context.h minimal, not #including other ruy headers.
Context is a powerful tool: it allows ruy to have some persistent resources managed by the user's code. One caveat is that since by design it's about exposing enough internals to the user to allow them to be managed by the user, it's a slippery slope towards exposing too much internals, inflating the user's translation units with ruy details, potentially running into more build issues and more refactoring friction in the future. We need to design this carefully to minimize how much internals are exposed, in particular, how much ruy internal code is #included by ruy public headers.

The basic first step here is a classical 'pimpl' or 'dpointer' idiom, the private members are moved to a CtxImpl class which Context needs to have a raw CtxImpl* pointer to, for which a forward-declaration is enough.

However, much internal ruy code needs to access that CtxImpl, and some of that code is templatized (front-end code in ruy.h, dispatch.h, etc) so needs to be in headers that are included into the user's translation units. So, we need to somehow let this ruy internal code access the CtxImpl without #including its definition.

This is achieved by an interface base class Ctx, which offers all the methods needed by internal ruy code, and is subclassed by CtxImpl. Thus the relationship between Context, Ctx and CtxImpl is:

Context has an owning raw pointer to a CtxImpl; CtxImpl subclasses Ctx.

The roles are:
Context is the user-facing class.
Ctx is what ruy's code uses.
CtxImpl is the Ctx subclass actually defining the data members. It is needed only in a few internal .cc files such as ctx.cc, context.cc.

PiperOrigin-RevId: 308181525