Split the storage of MulParams data members into 3 separate template specializations for the floating-point, raw integer and quantized cases.
The main goal was to reduce the worst-case sizeof(MulParams), i.e. the kMaxMulParamsSize constant used to set the size of the buffer in TrMulParams.

Sizes measured on x86-64 (would be the same on arm64):

AccumScalar,DstScalar     double,double   int32,int32   int32,int16
sizeof(MulParams) before  64              48            40
sizeof(MulParams) after   32              16            40

So kMaxMulParamsSize shrinks from 64 to 40.

The size 40 is not quite optimal. The layout is unchanged but there was no opportunity to pack data members better. The room for improvement is in how we are storing side by side mutually exclusive members. We could make them share a union instead. This would reduce the size from 40 to 32.

PiperOrigin-RevId: 318863047