Refactor Thread internals for clarity and efficiency.
On the clarity side, the thread main loop is now just:

    while (GetNewStateOtherThanReady() == State::HasWork) {
      RevertToReadyState();
    }

On the efficiency side:
* Locking and atomic ops have been reduced, we used to lock state_mutex_
  around the entire thread task execution,now we are only locking it
  anymore around notify/wait on the state_cond_ condition_variable, so
  this mutex is renamed state_cond_mutex_, which clarifies its purpose.
* We used to perform a redundant reload-acquire of the new state_ in
  the main thread loop.
* Some accesses are demoted to relaxed because they are already ordered
  by other release-acquire relationships.
* A notify_all becomes notify_one.
* Send all thread exit requests upfront so threads can exit in parallel.

A comment is added on Thread::task_ to explain the release-acquire
relationships making this all work.

Internal code is broken into functions that are only ever called
from the main thread, and functions that are only ever called from the
worker thread. That specialization
made further simplifications and performance gains obvious.

It was found by continuous integration that some TFLite users construct and destroy the context from two different threads, due to the use of reference-counting. That means that the notion of "main thread" is not that solid. Accordingly, instances of "main thread" in comments and identifiers have been rephrased as "outside thread" as opposed to worker thread.

Tested with TSan (also enabled on presubmits) so fairly confident that this
is correct.

PiperOrigin-RevId: 442697771