Free all memory before reallocating.
This should minimize the memory high-water-mark in applications. In the previous code, if there was a 1GB fallback block (and the main ptr_ was 0 bytes, say), then we would create a 1GB new allocation, and then free the 1GB fallback block, resulting in 2GB memory high water mark. In the new code, the memory high-water-mark will be only 1GB for that case.

In the best case this is expected to be a 2x savings on the memory high-water-mark. In practice, it may not reach 2x but it could be quite close -- it will be `(2 * (size of final steady state buffer allocation) - (next smaller total allocation size)) / (size of final steady state buffer allocation)`. In the case of "next smaller total allocation size" == 0 it achieves a 2x advantage. We expect that in practice the benefit will still be quite high, because this allocator specifically relies on reaching a steady state rapidly, and thus the steady state size must increase by very large jumps, meaning that (next smaller total allocation size) is likely much smaller than (size of final steady state buffer allocation).

PiperOrigin-RevId: 488405672