Disable AVX code paths on iOS simulator.
Fixes b/285639624.

The error there is an incorrect result from TFLite inference on iOS simulator when ruy's AVX code paths (which are written in intrinsics) are enabled.

Although I haven't looked at it in detail, the typical reason why this would happen --- given that this code has not seen any change in years and works as intended on a number of other platforms, including macOS/x86 --- would be a miscompilation by the iOS-simulator toolchain.

It is important to keep this code path enabled on macOS/x86 and only disable it on iOS simulator. We can't filter this by TARGET_OS_MAC as that token is defined in both toolchains. TARGET_IPHONE_SIMULATOR seems to be the appropriate differentiator.

PiperOrigin-RevId: 539714623