Add android x86-64 support
x86-64 is an officially supported arch on Android [1].
Android is a platform similar to Linux, and we can enable SIMD on it.

To do so, this CL extend `:x86_64_and_not_msvc` to not only linux but also android.
This will help apps on android that uses ruy to have SIMD speed up.

[1] https://developer.android.com/ndk/guides/abis#sa

PiperOrigin-RevId: 836149343